requirements.txt
Django==1.8.5
django-jsonfield==0.9.15
psycopg2==2.6.1
pytz==2015.6
wheel==0.24.0
